Add cqlTemplate to cassandra config
Add cql script and separate repos into event and image sessions

Add qualifier on CassandraOps

Fix cql script

Fix cql script

Fix cql script
