+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/dao/AuctionDaoJpa.java
@@ -40,8 +40,8 @@ import com.vmware.weathervane.auction.data.model.Keyword;
 import com.vmware.weathervane.auction.data.model.User;
 import com.vmware.weathervane.auction.data.model.Auction.AuctionState;
 import com.vmware.weathervane.auction.data.model.Item.ItemState;
-import com.vmware.weathervane.auction.data.repository.AttendanceRecordRepository;
-import com.vmware.weathervane.auction.data.repository.BidRepository;
+import com.vmware.weathervane.auction.data.repository.event.AttendanceRecordRepository;
+import com.vmware.weathervane.auction.data.repository.event.BidRepository;
 import com.vmware.weathervane.auction.util.FixedOffsetCalendarFactory;
 
 @Repository("auctionDao")
