+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/event/BidRepositoryCustom.java
@@ -13,15 +13,8 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.event;
 
-import java.util.UUID;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import com.vmware.weathervane.auction.data.model.NosqlBenchmarkInfo;
-
-@Repository
-public interface NosqlBenchmarkInfoRepository extends CrudRepository<NosqlBenchmarkInfo, UUID> {
+public interface BidRepositoryCustom {	
+	void deleteByItemId(Long itemId);
 }
