+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/event/NosqlBenchmarkInfoRepository.java
@@ -13,8 +13,15 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.event;
 
-public interface ImagePreviewRepositoryCustom {	
-	void deleteByKeyPreloaded(boolean preloaded);
+import java.util.UUID;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.vmware.weathervane.auction.data.model.NosqlBenchmarkInfo;
+
+@Repository
+public interface NosqlBenchmarkInfoRepository extends CrudRepository<NosqlBenchmarkInfo, UUID> {
 }
