+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/image/ImageFullRepositoryCustom.java
@@ -13,19 +13,8 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.image;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.cassandra.core.CassandraOperations;
-
-public class ImageFullRepositoryImpl implements ImageFullRepositoryCustom {
-
-	@Autowired
-	CassandraOperations cassandraOperations;
-	
-	@Override
-	public void deleteByPreloaded(boolean preloaded) {
-		String cql = "DELETE FROM image_full WHERE preloaded = " + preloaded + ";";
-		cassandraOperations.execute(cql);
-	}
+public interface ImageFullRepositoryCustom {	
+	void deleteByPreloaded(boolean preloaded);
 }
