+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/image/ImageFullRepositoryImpl.java
@@ -13,20 +13,21 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.image;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.cassandra.core.CassandraOperations;
 
-public class BidRepositoryImpl implements BidRepositoryCustom {
+public class ImageFullRepositoryImpl implements ImageFullRepositoryCustom {
 
 	@Autowired
+	@Qualifier("cassandraImageTemplate")
 	CassandraOperations cassandraOperations;
-
+	
 	@Override
-	public void deleteByItemId(Long itemId) {
-		String cql = "DELETE FROM bid_by_bidderid WHERE item_id = " + itemId + ";";
+	public void deleteByPreloaded(boolean preloaded) {
+		String cql = "DELETE FROM image_full WHERE preloaded = " + preloaded + ";";
 		cassandraOperations.execute(cql);
 	}
-
 }
