+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/image/ImageInfoRepositoryImpl.java
@@ -13,19 +13,21 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.image;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.cassandra.core.CassandraOperations;
 
-public class AttendanceRecordRepositoryImpl implements AttendanceRecordRepositoryCustom {
+public class ImageInfoRepositoryImpl implements ImageInfoRepositoryCustom {
 
 	@Autowired
+	@Qualifier("cassandraImageTemplate")
 	CassandraOperations cassandraOperations;
 	
 	@Override
-	public void deleteByAuctionId(Long auctionId) {
-		String cql = "DELETE FROM attendancerecord_by_userid WHERE auction_id = " + auctionId + ";";
+	public void deleteByPreloaded(boolean preloaded) {
+		String cql = "DELETE FROM image_info WHERE preloaded = " + preloaded + ";";
 		cassandraOperations.execute(cql);
 	}
 }
