+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/image/ImagePreviewRepositoryImpl.java
@@ -13,20 +13,21 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.image;
 
-import java.util.List;
-import java.util.UUID;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.cassandra.core.CassandraOperations;
 
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
+public class ImagePreviewRepositoryImpl implements ImagePreviewRepositoryCustom {
 
-import com.vmware.weathervane.auction.data.imageStore.model.ImageFull;
-import com.vmware.weathervane.auction.data.imageStore.model.ImageFull.ImageFullKey;
-
-@Repository
-public interface ImageFullRepository extends CrudRepository<ImageFull, ImageFullKey>, ImageFullRepositoryCustom {
-	List<ImageFull> findByKeyImageId(UUID imageid);
+	@Autowired
+	@Qualifier("cassandraImageTemplate")
+	CassandraOperations cassandraOperations;
 	
-	void deleteByKeyPreloaded(boolean preloaded);
+	@Override
+	public void deleteByKeyPreloaded(boolean preloaded) {
+		String cql = "DELETE FROM image_preview WHERE preloaded = " + preloaded + ";";
+		cassandraOperations.execute(cql);
+	}
 }
