+++ dbLoader/src/main/java/com/vmware/weathervane/auction/data/repository/image/ImageThumbnailRepositoryCustom.java
@@ -13,20 +13,9 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSE
 WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package com.vmware.weathervane.auction.data.repository;
+package com.vmware.weathervane.auction.data.repository.image;
 
-import java.util.List;
-import java.util.UUID;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import com.vmware.weathervane.auction.data.imageStore.model.ImageThumbnail;
-import com.vmware.weathervane.auction.data.imageStore.model.ImageThumbnail.ImageThumbnailKey;
-
-@Repository
-public interface ImageThumbnailRepository extends CrudRepository<ImageThumbnail, ImageThumbnailKey>, ImageThumbnailRepositoryCustom {
-	List<ImageThumbnail> findByKeyImageId(UUID imageid);
-	
+public interface ImageThumbnailRepositoryCustom {
 	void deleteByKeyPreloaded(boolean preloaded);
+
 }
